Alteração de double para BigDecimal

+++ src/main/java/com/pdv/lalapan/dto/ProdutoAtualizadoDTO.java
@@ -10,7 +10,7 @@ public record ProdutoAtualizadoDTO(
         String nome,
         BigDecimal preco,
         String codigo,
-        double quantidadeEstoque,
+        BigDecimal quantidadeEstoque,
         Unidade unidade,
         Categoria categoria
 ) {
